REF: Rename CXX files to match contents
