+++ include/noise.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <cuda.h>
 #include <cuda_runtime.h>
 #include <curand.h>
