+++ src/forward.cu
@@ -5,7 +5,7 @@
 
 #include "utils.h"
 #include "texture.h"
-#include "parameter_classes.h"
+#include "cfg.h"
 #include "log.h"
